+++ Tests/UnitTests/UnitTests.Shared/UnitTest1.cs
@@ -1,6 +1,3 @@
 namespace UnitTests.Shared;
 
-public class UnitTest1
-{
-
-}
+public class UnitTest1 { }