+++ Src/Library/changelog.md
@@ -1,6 +1,7 @@
 ### NEW
 - type safety for the shortcut http verb methods such as `Get()`, `Post()`, etc. [#info](https://fast-endpoints.com/docs/misc-conveniences#shorthand-route-configuration)
 - custom value parser registration at startup for any given type #285
+- specify whether to execute global pre/post processors before or after endpoint level processors #291
 
 ### IMPROVEMENTS
 - remove read-only properties from swagger request body #283